python/requirements2.txt
ConfigParser==6.0.0
pyserial==3.5
pyaml==20.4.0
pyyaml==6.0
pathlib==1.0.1